+++ Assets/Scripts/IDamagable.cs
@@ -1,7 +1,7 @@
 ﻿using UnityEngine;
 
 public interface IDamagable {
-   void TakeHit(float damage, RaycastHit hit);
+   void TakeHit(int damage, RaycastHit hit);
 
-    void TakeDamage(float damage);
+    void TakeDamage(int damage);
 }
